+++ lib/core/constants/app_constants.dart
@@ -3,4 +3,6 @@ class AppConstants {
 
   static const String appName = 'Fbro';
   static const String usersCollection = 'users';
+  static const String shiftsCollection = 'shifts';
+  static const String tasksCollection = 'tasks';
 }